evaluation/requirements.txt
requests
python-dotenv
tqdm 